requirements.txt
requests
pandas
yfinance
python-dotenv
psycopg2-binary
groq
matplotlib
seaborn
sqlalchemy